Moves blog-slice to components
