+++ site/compiler/people.test.js
@@ -34,7 +34,7 @@ describe('site/compiler', () => {
         badger: { [a.slug]: a },
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(14);
+      expect(routes.length).to.equal(15);
       expect(routes[10].filePath).to.equal('about-us/people/index.html');
       expect(routes[11].filePath).to.equal('about-us/people/category/engineering/index.html');
       expect(routes[12].filePath).to.equal('about-us/people/category/leadership/index.html');
@@ -70,7 +70,7 @@ describe('site/compiler', () => {
           categories,
         }, createStateNavigator());
 
-        expect(routes.length).to.equal(16);
+        expect(routes.length).to.equal(17);
         expect(routes[10].filePath).to.equal('about-us/people/index.html');
         expect(routes[11].filePath).to.equal('about-us/people/category/engineering/index.html');
         expect(routes[12].filePath).to.equal('about-us/people/category/leadership/index.html');
@@ -93,7 +93,7 @@ describe('site/compiler', () => {
           categories,
         }, createStateNavigator());
 
-        expect(routes.length).to.equal(13);
+        expect(routes.length).to.equal(14);
         expect(routes[10].filePath).to.equal('about-us/people/index.html');
         expect(routes[11].filePath).to.equal('about-us/people/category/ux-design/index.html');
       });
@@ -130,7 +130,7 @@ describe('site/compiler', () => {
           categories,
         }, createStateNavigator());
 
-        expect(routes.length).to.equal(36);
+        expect(routes.length).to.equal(37);
         expect(routes[10].filePath).to.equal('about-us/people/index.html');
         expect(routes[11].filePath).to.equal('about-us/people/category/everyone/page-2/index.html');
         expect(routes[12].filePath).to.equal('about-us/people/category/engineering/index.html');
@@ -160,7 +160,7 @@ describe('site/compiler', () => {
           categories,
         }, createStateNavigator());
 
-        expect(routes.length).to.equal(33);
+        expect(routes.length).to.equal(34);
         expect(routes[10].filePath).to.equal('about-us/people/index.html');
         expect(routes[11].filePath).to.equal('about-us/people/category/everyone/page-2/index.html');
         expect(routes[12].filePath).to.equal('about-us/people/category/engineering/index.html');
@@ -189,7 +189,7 @@ describe('site/compiler', () => {
         categories,
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(31);
+      expect(routes.length).to.equal(32);
       expect(routes[10].filePath).to.equal('about-us/people/index.html');
       expect(routes[11].filePath).to.equal('about-us/people/category/engineering/index.html');
     });
@@ -219,7 +219,7 @@ describe('site/compiler', () => {
         categories,
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(60);
+      expect(routes.length).to.equal(61);
       expect(routes[10].filePath).to.equal('about-us/people/index.html');
       expect(routes[11].filePath).to.equal('about-us/people/category/everyone/page-2/index.html');
       expect(routes[12].filePath).to.equal('about-us/people/category/everyone/page-3/index.html');
@@ -265,7 +265,7 @@ describe('site/compiler', () => {
         categories,
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(59);
+      expect(routes.length).to.equal(60);
       expect(routes[10].filePath).to.equal('about-us/people/index.html');
       expect(routes[11].filePath).to.equal('about-us/people/category/everyone/page-2/index.html');
       expect(routes[12].filePath).to.equal('about-us/people/category/everyone/page-3/index.html');
@@ -294,7 +294,7 @@ describe('site/compiler', () => {
         badger: { [a.slug]: a },
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(13);
+      expect(routes.length).to.equal(14);
       expect(routes[12].filePath).to.equal('about-us/people/alex/index.html');
     });
   });
@@ -325,7 +325,7 @@ describe('site/compiler', () => {
         },
       }, createStateNavigator());
 
-      expect(routes.length).to.equal(14);
+      expect(routes.length).to.equal(15);
       expect(routes[12].filePath).to.equal('about-us/people/alex/index.html');
       expect(routes[13].filePath).to.equal('about-us/people/sari/index.html');
     });
