+++ site/routes/definitions.js
@@ -107,6 +107,9 @@ export const routeDefinitions: Array<RouteDefinition> = [
     title: 'Technology',
     key: 'technology',
     route: 'technology',
+    stateToProps: ({ triedAndTestedBlogPosts }) => ({
+      triedAndTestedBlogPosts,
+    }),
   },
   {
     title: 'Not found',
